[Feat/#72] 긴급: memberResponse 추가에 따른, api response 추가 및 appdatastore 개편

+++ data/src/main/java/com/umc/data/response/challenger/ChallengerPointResponse.kt
@@ -3,7 +3,7 @@ package com.umc.data.response.challenger
 import com.google.gson.annotations.SerializedName
 
 data class ChallengerPointResponse(
-    @SerializedName("id") val id: Int? = null,
+    @SerializedName("id") val id: Long? = null,
     @SerializedName("pointType") val pointType: String? = null,
     @SerializedName("point") val point: Double? = null,
     @SerializedName("description") val description: String? = null
